requirements.txt
opencv-python
mysql-connector-python
pandas
torch
torchvision
ultralytics